+++ treehugger/cli/__init__.py
@@ -3,19 +3,17 @@ from __future__ import absolute_import, division, print_function, unicode_litera
 
 import argparse
 import os
-import subprocess
 import sys
-import tempfile
 
 from six.moves import shlex_quote
 
 from .decrypt_file import decrypt_file
+from .edit import edit
+from .encrypt_file import encrypt_file
 from .parser import parser, subparsers
-from .utils import load_and_decrypt_file, load_and_encrypt_file
 from .. import yaml
 from ..data import EnvironmentDict
 from ..ec2 import load_user_data_as_yaml_or_die
-from ..os_ext import patch_umask
 
 
 def main(args_list=None):
@@ -26,64 +24,11 @@ def main(args_list=None):
     func(args)
 
 
-command_funcs = {}
-
-
-command_funcs['decrypt-file'] = decrypt_file
-
-
-encrypt_file_parser = subparsers.add_parser(
-    'encrypt-file',
-    description='Encrypt a Treehugger YAML file in-place.'
-)
-encrypt_file_parser.add_argument('filename', type=str, help='The path to the file to encrypt')
-
-
-def encrypt_file(args):
-    filename = args.filename
-
-    new_data = load_and_encrypt_file(filename)
-    yaml.save_file(filename, new_data)
-
-    print('Successfully encrypted')
-
-
-command_funcs['encrypt-file'] = encrypt_file
-
-edit_parser = subparsers.add_parser(
-    'edit',
-    description='Decrypt a treehugger YAML file temporarily, edit it with $EDITOR, then re-encrypt it.',
-)
-edit_parser.add_argument('filename', type=str, help='The path to the file to edit')
-
-
-def edit(args):
-    filename = args.filename
-
-    data = load_and_decrypt_file(filename)
-
-    temp_fp = tempfile.NamedTemporaryFile(suffix='.yml', mode='w', delete=False)
-    yaml.save_fp(temp_fp, data)
-    temp_fp.flush()
-    temp_fp.close()
-
-    with patch_umask(0x077):  # to stop $EDITOR from creating world-readable temp files
-        editor = os.environ.get('EDITOR', 'nano')
-        retcode = subprocess.call([editor, temp_fp.name])
-
-    if retcode != 0:
-        os.unlink(temp_fp.name)
-        print('Editor failed with return code {}'.format(retcode), file=sys.stderr)
-        raise SystemExit(1)
-
-    new_data = load_and_encrypt_file(temp_fp.name)
-    os.unlink(temp_fp.name)
-
-    yaml.save_file(filename, new_data)
-    print('Successfully edited file')
-
-
-command_funcs['edit'] = edit
+command_funcs = {
+    'decrypt-file': decrypt_file,
+    'encrypt-file': encrypt_file,
+    'edit': edit,
+}
 
 
 exec_parser = subparsers.add_parser(
